feito alguns ajustes
