+++ ds2m/caio_palermo_lemos/index.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta name="description" content="WhatsApp Site" />
     <script
       src="https://kit.fontawesome.com/c0e5486155.js"
       crossorigin="anonymous"
@@ -74,7 +75,11 @@
         </div>
         <div class="chats__filter">
           <div class="filter__input-container">
-            <input type="text" class="filter__input" placeholder="Pesquisar" />
+            <input
+              type="search"
+              class="filter__input"
+              placeholder="Pesquisar"
+            />
             <i
               class="fa-solid fa-magnifying-glass"
               id="filter__icon-search"
