+++ ds2m/caio_palermo_lemos/js/app.js
@@ -1,185 +1,133 @@
 "use strict";
 
 import { contatos } from "./contatos.js";
-import { createChat } from "./create-chat.js";
-import { showContactMessages } from "./show-contact-messages.js";
 import { changeElements } from "./change-elements.js";
-import { createMessagesReceiver } from "./create-messages-receiver.js";
 import { returnIcon } from "./return.js";
+import { chargeContacts } from "./charge-contacts.js";
+import { chargeProfile } from "./charge-profile.js";
 
-const chargeContacts = function () {
-  const chatsContainer = document.querySelector(".chats__chats-container");
-  const chats = contatos.map(createChat);
-
-  chatsContainer.replaceChildren(...chats);
-};
-
-const chargeProfile = function () {
-  const handleClick = function (event) {
-    let text = event.currentTarget.innerText.split(`\n`)[0];
-
-    const messagesReceiver = document.querySelector(".messages__receiver");
-    const profile = contatos.map(createMessagesReceiver);
-
-    const indice = contatos.findIndex((contato) => {
-      return contato.name == text;
-    });
-
-    messagesReceiver.replaceChildren(profile[indice]);
-  };
-
-  const chats = document.querySelectorAll(".chat");
-  chats.forEach((chat) => {
-    chat.addEventListener("click", handleClick);
-  });
-};
-
-chargeContacts();
-chargeProfile();
+chargeContacts(contatos);
+chargeProfile(contatos);
 changeElements();
+returnIcon();
 
-// AQUI TA DANDO A MERDA
-const fazerTudo = function (message) {
-  // function convertTZ(date, tzString) {
-  //   return new Date(
-  //     (typeof date === "string" ? new Date(date) : date).toLocaleString(
-  //       "en-US",
-  //       {
-  //         timeZone: tzString,
-  //       }
-  //     )
-  //   );
-  // }
+function convertTZ(date, tzString) {
+  return new Date(
+    (typeof date === "string" ? new Date(date) : date).toLocaleString("en-US", {
+      timeZone: tzString,
+    })
+  );
+}
 
-  const fazerLoop = function (indice) {
-    for (let i = 0; contatos[indice].messages.length > i; i++) {
-      const createMessages = function () {
-        message = contatos[indice].messages[i];
+const fazerLoop = function (indice) {
+  for (let i = 0; contatos[indice].messages.length > i; i++) {
+    const createMessages = function (message) {
+      message = contatos[indice].messages[i];
 
-        let messagesConversations = document.createElement("div");
-        messagesConversations.classList.add("messages__conversations");
+      let messagesConversations = document.createElement("div");
+      messagesConversations.classList.add("messages__conversations");
 
-        let messageSenderContainer = document.createElement("div");
-        messageSenderContainer.classList.add("message-sender-container");
+      let messageSenderContainer = document.createElement("div");
+      messageSenderContainer.classList.add("message-sender-container");
 
-        let messageSender = document.createElement("div");
-        messageSender.classList.add("message__sender");
-        messageSender.classList.add("tri-right");
-        messageSender.classList.add("left-top");
+      let messageSender = document.createElement("div");
+      messageSender.classList.add("message__sender");
+      messageSender.classList.add("tri-right");
+      messageSender.classList.add("left-top");
 
-        let messageReceiver = document.createElement("div");
-        messageReceiver.classList.add("message__receiver");
-        messageReceiver.classList.add("tri-right");
-        messageReceiver.classList.add("right-top");
+      let messageReceiver = document.createElement("div");
+      messageReceiver.classList.add("message__receiver");
+      messageReceiver.classList.add("tri-right");
+      messageReceiver.classList.add("right-top");
 
-        let messageContent = document.createElement("p");
+      let messageContent = document.createElement("p");
+      messageContent.classList.add("message__content");
+      messageContent.textContent = message.content;
+
+      let messageTime = document.createElement("span");
+      messageTime.classList.add("message__time");
+
+      let messageReceiverContainer = document.createElement("div");
+      messageReceiverContainer.classList.add("message-receiver-container");
+
+      if (message.sender == "me") {
+        if (message.hasOwnProperty("timestamp")) {
+          const dataConvertida = convertTZ(
+            message.timestamp,
+            "America/Sao_Paulo"
+          );
+          message.timestamp = `${dataConvertida.getHours()}:${dataConvertida.getMinutes()}`;
+          messageTime.textContent = message.timestamp;
+        }
+        if (message.hasOwnProperty("time")) {
+          messageTime.textContent = message.time;
+        }
+
         messageContent.classList.add("message__content");
         messageContent.textContent = message.content;
 
-        let messageTime = document.createElement("span");
         messageTime.classList.add("message__time");
 
-        // const dataConvertida = convertTZ(
-        //   message.timestamp,
-        //   "America/Sao_Paulo"
-        // );
+        messageSender.replaceChildren(messageContent, messageTime);
 
-        let messageReceiverContainer = document.createElement("div");
-        messageReceiverContainer.classList.add("message-receiver-container");
-
-        if (message.sender == "me") {
-          messageContent.classList.add("message__content");
-          messageContent.textContent = message.content;
-
-          messageTime.classList.add("message__time");
+        messageReceiverContainer.classList.add("none");
+      } else {
+        if (message.hasOwnProperty("timestamp")) {
+          const dataConvertida = convertTZ(
+            message.timestamp,
+            "America/Sao_Paulo"
+          );
+          message.timestamp = `${dataConvertida.getHours()}:${dataConvertida.getMinutes()}`;
+          messageTime.textContent = message.timestamp;
+        }
+        if (message.hasOwnProperty("time")) {
           messageTime.textContent = message.time;
-
-          messageSender.replaceChildren(messageContent, messageTime);
-
-          messageReceiverContainer.classList.add("none");
-        } else {
-          messageContent.classList.add("message__content");
-          messageContent.textContent = message.content;
-
-          // if (message.hasOwnProperty(time)) {
-          //   messageTime.classList.add("message__time");
-          //   messageTime.textContent = `${dataConvertida.getHours()}:${dataConvertida.getMinutes()}`;
-          // }
-          // if (message.hasOwnProperty(timestamp)) {
-          //   messageTime.textContent = message.time;
-          // }
-
-          messageTime.classList.add("message__time");
-          messageTime.textContent = message.time;
-
-          messageReceiver.replaceChildren(messageContent, messageTime);
-
-          messageSenderContainer.classList.add("none");
         }
 
-        messageSenderContainer.replaceChildren(messageSender);
-        messageReceiverContainer.replaceChildren(messageReceiver);
+        messageContent.classList.add("message__content");
+        messageContent.textContent = message.content;
 
-        messagesConversations.replaceChildren(
-          messageSenderContainer,
-          messageReceiverContainer
-        );
+        messageTime.classList.add("message__time");
 
-        return messagesConversations;
-      };
+        messageReceiver.replaceChildren(messageContent, messageTime);
+        messageSenderContainer.classList.add("none");
+      }
 
-      const messagesConversationsContainer = document.querySelector(
-        ".messages__conversations-container"
+      messageSenderContainer.replaceChildren(messageSender);
+      messageReceiverContainer.replaceChildren(messageReceiver);
+
+      messagesConversations.replaceChildren(
+        messageSenderContainer,
+        messageReceiverContainer
       );
 
-      messagesConversationsContainer.append(createMessages());
-    }
-  };
-  const handleClick = function (event) {
+      return messagesConversations;
+    };
+
     const messagesConversationsContainer = document.querySelector(
       ".messages__conversations-container"
     );
-    let text = event.currentTarget.innerText.split(`\n`)[0];
-    console.log(text);
 
-    let indice = 0;
-    indice = contatos.findIndex((contato) => {
-      return contato.name == text;
-    });
-    console.log(indice);
+    messagesConversationsContainer.append(createMessages());
+  }
+};
+const handleClick = function (event) {
+  const messagesConversationsContainer = document.querySelector(
+    ".messages__conversations-container"
+  );
+  let text = event.currentTarget.innerText.split(`\n`)[0];
 
-    messagesConversationsContainer.innerHTML = "";
-
-    showContactMessages(contatos);
-
-    fazerLoop(indice);
-  };
-
-  const chats = document.querySelectorAll(".chat");
-  chats.forEach((chat) => {
-    chat.addEventListener("click", handleClick);
+  let indice = 0;
+  indice = contatos.findIndex((contato) => {
+    return contato.name == text;
   });
+
+  messagesConversationsContainer.innerHTML = "";
+
+  fazerLoop(indice);
 };
 
-fazerTudo(contatos);
-
-returnIcon();
-// contatos.forEach((contato) => {
-//   contato.messages.forEach((message) => {
-//     console.log(message.timestamp);
-//   });
-// });
-
-// console.log(`${convertedDate.getHours()}:${convertedDate.getMinutes()}`); // 17
-
-//BARRA DE PESQUISAR
-const buscar = function (contatos) {
-  let nome = "Ana";
-  const chats = document.querySelectorAll(".chat__name");
-
-  contatos.forEach((item) => {
-    if (item.name.startsWith("Ana")) console.log(item.name);
-  });
-};
-
-buscar(contatos);
+const chats = document.querySelectorAll(".chat");
+chats.forEach((chat) => {
+  chat.addEventListener("click", handleClick);
+});
