+++ ds2m/caio_palermo_lemos/js/charge-contacts.js
@@ -1,0 +1,8 @@
+import { createChat } from "./create-chat.js";
+
+export const chargeContacts = function (contatos) {
+  const chatsContainer = document.querySelector(".chats__chats-container");
+  const chats = contatos.map(createChat);
+
+  chatsContainer.replaceChildren(...chats);
+};
