+++ ds2m/caio_palermo_lemos/js/search-contact.js
@@ -1,0 +1,30 @@
+//BARRA DE PESQUISAR
+// const buscar = function () {
+//   const filterInput = document.querySelector(".filter__input");
+//   const chats = document.querySelectorAll(".chat");
+
+//   const handleKey = function (event) {
+//     console.log(event.key);
+//     filterInput.value = "";
+//     if (filterInput.value != "") {
+//       contatos.forEach((contato) => {
+//         filterInput.value = "Mark";
+//         if (contato.name.startsWith(filterInput.value)) {
+//           chats.forEach((item) => {
+//             let nome = item.innerText.split(`\n`)[0];
+//             item.classList.add("none");
+
+//             if (nome == contato.name) {
+//               console.log("O nome é: " + nome);
+//               item.classList.remove("none");
+//             }
+//           });
+//         }
+//       });
+//     }
+//   };
+
+//   filterInput.addEventListener("keypress", handleKey);
+// };
+
+// buscar();
