--- ds2m/caio_palermo_lemos/js/show-contact-messages.js
@@ -1,22 +1,0 @@
-export const showContactMessages = function (contatos) {
-  const selectedContact = function () {
-    let contatoSelecionado;
-    let arrayContato;
-
-    contatos.forEach((contato) => {
-      const messages = contato.messages;
-      messages.filter((message) => {
-        if (message.sender == contatoSelecionado) {
-          arrayContato = messages;
-          console.log(arrayContato);
-          return arrayContato;
-        }
-      });
-    });
-  };
-
-  const chats = document.querySelectorAll(".chat");
-  chats.forEach((chat) => {
-    chat.addEventListener("click", selectedContact);
-  });
-};
